pyproject.toml
[project]
name = "testgen"
version = "0.1.0"
description = "Автоматическая генерация unit-тестов для C++ проектов с использованием LLM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.8.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.3",
    "libclang>=16.0.0",      # ← добавили
]

[project.scripts]
testgen = "testgen.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]
include = ["testgen*"]
namespaces = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "SIM", "UP", "RUF"]
ignore = ["E501"]
fixable = ["ALL"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]